modifications to a11y section
make note that mention of aria-controls should be considered TBD. Adds mention of aria-details as that may be more likely to be used, as it is used with popover now. Added in mention of that to the section talking about popovers.  

Fixed a number of typos.

Removed mention of aria-haspopup=dialog.  It's use is debatable in its usefulness.